refactor(auth): rename _authenticated module to authenticated
The underscore-prefixed module name was inconsistent with the rest of
services/ (base classes use plain names like base_broker, base_executor). No
behavior change; pure rename of services/_authenticated.py -> authenticated.py
and its imports.
