+++ libs/aegra-api/src/aegra_api/services/authenticated.py
