+++ libs/aegra-api/tests/integration/test_services/test_assistant_service_db.py
@@ -210,12 +210,12 @@ class TestAssistantServiceDatabase:
         mock_result.all.return_value = []
         assistant_service.session.scalars.return_value = mock_result
 
-        from aegra_api.services import _authenticated
+        from aegra_api.services import authenticated
 
         async def handler_filter(ctx, value):
             return {"metadata": {"tenant": "x"}}
 
-        monkeypatch.setattr(_authenticated, "handle_event", handler_filter)
+        monkeypatch.setattr(authenticated, "handle_event", handler_filter)
 
         result = await assistant_service.list_assistants()
