+++ libs/aegra-api/tests/unit/test_services/test_authenticated.py
@@ -5,7 +5,7 @@ from unittest.mock import AsyncMock, patch
 import pytest
 
 from aegra_api.models.auth import User
-from aegra_api.services._authenticated import Authenticated
+from aegra_api.services.authenticated import Authenticated
 
 
 class _Service(Authenticated):
@@ -21,7 +21,7 @@ class TestDispatch:
     @pytest.mark.asyncio
     async def test_builds_context_with_resource_and_action(self, service: _Service) -> None:
         with patch(
-            "aegra_api.services._authenticated.handle_event",
+            "aegra_api.services.authenticated.handle_event",
             new=AsyncMock(return_value=None),
         ) as handle_event:
             await service._dispatch("read", {"thread_id": "t-1"})
@@ -35,7 +35,7 @@ class TestDispatch:
     async def test_returns_handler_filters(self, service: _Service) -> None:
         filters = {"user_id": "user-123"}
         with patch(
-            "aegra_api.services._authenticated.handle_event",
+            "aegra_api.services.authenticated.handle_event",
             new=AsyncMock(return_value=filters),
         ):
             result = await service._dispatch("search", {})
@@ -48,7 +48,7 @@ class TestDispatch:
 
         with (
             patch(
-                "aegra_api.services._authenticated.handle_event",
+                "aegra_api.services.authenticated.handle_event",
                 new=AsyncMock(side_effect=HTTPException(status_code=403, detail="Forbidden")),
             ),
             pytest.raises(HTTPException) as exc_info,
